Refactor redirect page and update dependencies
- Separate redirect logic into a client component
- Add @emotion/is-prop-valid dependency
- Update package dependencies to latest versions
- Modify homepage testimonial text
- Add unoptimized prop to Image component
