+++ package.json
@@ -11,6 +11,7 @@
     "biome:lint": "biome lint --write ./src"
   },
   "dependencies": {
+    "@emotion/is-prop-valid": "^1.3.1",
     "framer-motion": "^12.4.10",
     "next": "15.2.0",
     "react": "^19.0.0",
